Divide emissions form into smaller components
